cadaluan/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
pillow==10.2.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
